refractor: useRepoDetails hook
